food-assistant-backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
bcrypt==4.2.0
PyJWT==2.9.0
pymongo==4.8.0
redis==5.0.7
python-dotenv==1.0.1
werkzeug==3.0.3